chore: change the structure of sections, subsections and folder structure
