+++ libs/application/templates/rental-agreement/src/forms/draftForm/costsSection/index.ts
@@ -1,0 +1,15 @@
+import { buildSection } from '@island.is/application/core'
+import { rentalAmountSubsection } from './rentalAmountSubsection'
+import { securityDepositSubsection } from './securityDepositSubsection'
+import { otherFeesSubsection } from './otherFeesSubsection'
+import * as m from '../../../lib/messages'
+
+export const costsSection = buildSection({
+  id: 'costsSection',
+  title: m.costsMessages.sectionTitle,
+  children: [
+    rentalAmountSubsection,
+    securityDepositSubsection,
+    otherFeesSubsection,
+  ],
+})
