+++ libs/application/templates/rental-agreement/src/forms/draftForm/costsSection/rentalAmountSubsection.ts
@@ -31,7 +31,7 @@ import {
   getRentalAmountPaymentDateOptions,
 } from '../../../utils/options'
 
-export const RentalPeriodAmount = buildSubSection({
+export const rentalAmountSubsection = buildSubSection({
   id: Routes.RENTALAMOUNT,
   title: m.rentalAmount.subSectionName,
   children: [
