+++ libs/application/templates/rental-agreement/src/forms/draftForm/partiesSection/index.ts
@@ -1,9 +1,44 @@
-import { buildSection } from '@island.is/application/core'
-import { partiesSubsection } from './partiesSubsection'
+import {
+  buildMultiField,
+  buildSection,
+  buildTableRepeaterField,
+} from '@island.is/application/core'
 import * as m from '../../../lib/messages'
+import {
+  applicantTableConfig,
+  landlordTableConfig,
+  applicantTableFields,
+  landLordInfoTableFields,
+} from '../../../utils/utils'
+import { Routes } from '../../../utils/enums'
 
-export const PartiesSection = buildSection({
-  id: 'partiesSection',
+export const partiesSection = buildSection({
+  id: Routes.PARTIESINFORMATION,
   title: m.partiesDetails.subSectionName,
-  children: [partiesSubsection],
+  children: [
+    buildMultiField({
+      id: Routes.PARTIESINFORMATION,
+      title: m.partiesDetails.multiFieldTitle,
+      description: m.partiesDetails.multiFieldDescription,
+      children: [
+        buildTableRepeaterField({
+          id: 'parties.landlordInfo.table',
+          title: m.partiesDetails.landlordTableTitle,
+          editField: true,
+          marginTop: 1,
+          maxRows: 10,
+          fields: landLordInfoTableFields,
+          table: landlordTableConfig,
+        }),
+        buildTableRepeaterField({
+          id: 'parties.tenantInfo.table',
+          title: m.partiesDetails.tenantTableTitle,
+          editField: true,
+          maxRows: 10,
+          fields: applicantTableFields,
+          table: applicantTableConfig,
+        }),
+      ],
+    }),
+  ],
 })
