+++ libs/application/templates/rental-agreement/src/lib/messages/costsMessages.ts
@@ -1,0 +1,9 @@
+import { defineMessages } from 'react-intl'
+
+export const costsMessages = defineMessages({
+  sectionTitle: {
+    id: 'ra.application:costsMessages.sectionTitle',
+    defaultMessage: 'Kostnaður',
+    description: 'Section title for the costs section',
+  },
+})
