+++ libs/application/templates/rental-agreement/src/lib/messages/index.ts
@@ -14,3 +14,4 @@ export * from './reviewInfo'
 export * from './signing'
 export * from './summary'
 export * from './misc'
+export * from './costsMessages'
